fix: Path Parm으로 전달될 수 있는 값의 범위를 고려하여 UUID max 값을 integer max값을 기준으로 설정

+++ app/src/main/java/com/codesoom/assignment/Task.java
@@ -42,6 +42,6 @@ public class Task implements DefaultTask, Comparable<Task> {
     }
 
     private Long generatedId() {
-        return UUID.randomUUID().getMostSignificantBits() & Long.MAX_VALUE;
+        return UUID.randomUUID().getMostSignificantBits() & Integer.MAX_VALUE;
     }
 }
